Renamed parse to top_command
